fix: execute all tests when using all parameter & typo
